pyproject.toml
[project]
name = "epub2speech"
version = "0.1.0"
description = ""
authors = [
    {name = "TaoZeyu",email = "i@taozeyu.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ebooklib (>=0.19,<0.20)",
    "azure-cognitiveservices-speech (>=1.46.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
